requirements.txt
h2o_wave==0.20.0
loguru
matplotlib
toml
pandas
pandasql
psycopg2-binary
pyjwt
isort==5.7.0
black==20.8b1
timestring
numpy
nltk
gensim
spacy
pyLDAvis
datatable
filelock
http://h2o-public-test-data.s3.amazonaws.com/e2e-testing/appstore-0.0.1-py3-none-any.whl
https://enterprise-steam.s3.amazonaws.com/release/1.8.9/python/h2osteam-1.8.9-py2.py3-none-any.whl
